header moved to inc dir and makefile update
