+++ inc/minishell.h
@@ -6,14 +6,14 @@
 /*   By: gbuczyns <gbuczyns@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/20 21:48:08 by marvin            #+#    #+#             */
-/*   Updated: 2024/08/21 22:21:36 by gbuczyns         ###   ########.fr       */
+/*   Updated: 2024/08/22 21:11:32 by gbuczyns         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef MINISHELL_H
 # define MINISHELL_H
 
-# include "./lib_ft/libft.h"
+# include "../lib_ft/libft.h"
 # include <readline/history.h>
 # include <readline/readline.h>
 # include <stdbool.h>
@@ -28,7 +28,7 @@
 // DRBUG_FIELD
 // DELETE ME befor release
 //========================================================================================//
-# include "./lib_ft_printf/ft_printf.h"
+# include "../lib_ft_printf/ft_printf.h"
 //========================================================================================//
 typedef struct s_env
 {
