+++ src/validate_input.c
